+++ src/library_pthread.js
@@ -428,9 +428,6 @@ var LibraryPThread = {
 #if USE_OFFSET_CONVERTER
         'wasmOffsetConverter': wasmOffsetConverter,
 #endif
-#if !MINIMAL_RUNTIME
-        'DYNAMIC_BASE': DYNAMIC_BASE
-#endif
       });
     },